+++ src/stan/mcmc/hmc/hamiltonians/diag_e_point.hpp
@@ -28,7 +28,7 @@ namespace stan {
         if(!o) return;
         *o << "# Diagonal elements of inverse mass matrix:" << std::endl;
         *o << "# " << mInv(0) << std::flush;
-        for(size_t i = 1; i < mInv.size(); ++i)
+        for (int i = 1; i < mInv.size(); ++i)
           *o << ", " << mInv(i) << std::flush;
         *o << std::endl;
       };
